feat(ansible): add playbook for configuring app and db servers
This commit adds a new playbook file for configuring the app server and the db server in an Ansible lab environment. The playbook includes tasks such as configuring hosts, installing software packages (Java, Maven, Git), creating directories, cloning a Git repository, generating application packages, configuring systemd service, and starting the notes service.

The commit also introduces two new roles, "configuracao-default-so" and "geerlingguy.mysql," which are used to perform system configuration and install MySQL, respectively.

This update aims to automate the setup process of the app and db servers, making it easier to provision and manage the infrastructure.

The playbook and associated files can be found in the `lab-ansible/control-node/playbooks` directory.